Fixes #2
While the disclaimer for the package states that this package was written for educational use only. #2 pointed out that the reported support for reduced and/or amibiguous alphabets was not true.

After looking into this, it was determined that none of test cases tested this assumption. The underlying code was not fully fleshed out.

After a semi-significant rewrite. this should allow for the analysis of reduced and/or ambiguous alphabets. The **main caveat** being that *all ambiguous characters are degenerated  all possible canonical letters*. That is, in the case of "V" in the DNA ambiguous alphabet, its counts are divided amongst the "A", "C", and "G" columns.