Fixes #5
This biggest issue here is that the vanilla notebooks treated SVGs as images, such that the IDs used in their generation were not treated as unique. Therefore, when painted onto the DOM, since SVGs could have non-unique IDs, they would collide with each other. This isn't really seen in jupyterlab because of how it does rendering, but this issues was originally seen in https://github.com/ipython/ipython/issues/1866/. I fixed this by quickly parsing the SVG data and inserting the original hash of the data into the unique ids. Kind of a pain, but it fixes this problem